Move SearchFormSortByPortal into the SearchForm directory
